Improve performance by doing IO in blocks
On adapters that support it (future 3.4.x firmwares) a list of
transactions (read/write IO on controllers) can be sent through
a single USB request. This reduces the number of back and forths
(currently one USB request per controller read or write) and
therefore blocks the emulator for a shorter period.