Support repeated pb_scanControllers() calls
In project64, InitiateControllers is called each time a new
game is opened. Before this commit, this caused a bug where
the adapters connected to the system were appended to the list,
even if there was already open handles on them. (This means that
after 3 game restarts, a single-player controller would be controller
players 1 to 3...)