Fix controller swapping for some games.
In some games, once the timeout bit (0x80) in the R byte gets set
(when disconnecting a controller to swap it, for instance), it remains
set (i.e. the game does not clear it before each poll) until a transfer
succeeds.

This is not incorrect. This bit should be ignored when receiving, it is not part
of the length. i.e. A R value of 0x84 is valid and means receive=4.

This Battle Tanx workaround would ignore any value above 64, therefore
all values where bit 7 (0x80) is set, leading to no more transfers being
done, and therefore bit 7 never cleared by a succeeding transfer.

The workaround is now more specific.